run_scons: call scons from a bash script, for signals
and so, to call scons, we have a bash wrapper, run_scons.sh, which
creates two wrappers, valeNsconsM.bat and valeNsconsM.sh, and calls
the .bat, which then calls the .sh with bash, which then calls scons
with python.

The .bat layer is needed because we need to set up the Visual Studio
environment through vcvarsall.bat, which only works from within a .bat
file. Then, the .sh layer is needed so that UNIX signals (e.g. CTRL+C)
can be caught while running scons (which would be impossible if run
directly from the .bat)

Enjoy!