Issue #28:  Calc for F*
Replace CaclOpOpt with CalcOp for both CalcContents and CALC.  Change the semantics of the CalcOp in CALC  (e.g. calc == { )to match F*, which indicate the final relation between the first value and final value (whereas previously, it is used as user provided defaultOp as in Dafny. Since CalOp is required in CalcContent, we no longer need user provided defaultOp, therefore, it is omitted when emitted as dafny text.)