Have {:transform} lemma gen higher priority than quick. See msg.
If we have it as a lower priority, then we are unable to use {:quick}
for the untransformed code's proof. This isn't ideal. Instead, we'd
like to be able to use {:quick} for the untransformed code, and then
use the specialized lemma proof for the transformed code. This change
allows this by correctly handling priorities.